Splited js code into js  modules
